refactor table implementation to separate component
